requirements.txt
openai-whisper
keybert
nltk
transformers
torch
librosa
pydub
sentence-transformers
